requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
contourpy==1.1.1
cycler==0.11.0
filelock==3.9.0
fonttools==4.42.1
fsspec==2023.9.2
huggingface-hub==0.17.3
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.2
matplotlib==3.8.0
mpmath==1.3.0
networkx==3.0
numpy==1.24.1
packaging==23.1
Pillow==9.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
PyYAML==6.0.1
regex==2023.8.8
requests==2.28.1
safetensors==0.3.3
six==1.16.0
sympy==1.12
timm==0.9.7
tokenizers==0.13.3
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
tqdm==4.66.1
transformers==4.33.2
typing_extensions==4.4.0
urllib3==1.26.13
